Main screen styling except 3d effect and answer fill.
