+++ elements/reigns/src/Header.tsx
@@ -3,16 +3,13 @@ import { GameDefinition } from "./features/game/types";
 
 export const Header = ({
   definition,
-  stats,
-  round,
+  stats
 }: {
   definition: GameDefinition;
   stats: number[];
-  round: number;
 }) => {
   return (
     <div className="block header">
-      <div className="header__side" />
       <div className="meters">
         {definition.stats.map((stat, ix) => (
           <Meter
@@ -22,9 +19,6 @@ export const Header = ({
             name={stat.name}
           />
         ))}
-      </div>
-      <div className="header__side round">
-        {definition.roundName} {round}
       </div>
     </div>
   );
